Migrate LFS core logic to LFS module.
Routers are moved to `/git/lfs/http.rs`.
