+++ src/git/lfs/mod.rs
@@ -1,1 +1,2 @@
+pub mod http;
 pub mod structs;
