+++ src/git/object/base/mod.rs
@@ -45,5 +45,3 @@ impl ObjectClass {
         meta
     }
 }
-
-
